Change MTU Size to 517

+++ features/FEATURE_BLE/targets/TARGET_NORDIC/TARGET_NORDIC_SOFTDEVICE/TARGET_NRF52/source/btle/btle.cpp
@@ -352,12 +352,12 @@ void btle_handler(const ble_evt_t *p_ble_evt)
         case BLE_GAP_EVT_DATA_LENGTH_UPDATE_REQUEST:
         {
             ble_gap_evt_t       const * p_gap_evt = &p_ble_evt->evt.gap_evt;
-            uint8_t const data_length_peer =
+            uint16_t const data_length_peer =
                 p_gap_evt->params.data_length_update_request.peer_params.max_tx_octets;
 
-            const uint8_t max_data_length = NRF_SDH_BLE_GATT_MAX_MTU_SIZE + 4 /* L2CAP header size */;
+            const uint16_t max_data_length = NRF_SDH_BLE_GATT_MAX_MTU_SIZE + 4 /* L2CAP header size */;
 
-            uint8_t const data_length = MIN(max_data_length, data_length_peer);
+            uint16_t const data_length = MIN(max_data_length, data_length_peer);
 
             ble_gap_data_length_params_t const dlp =
             {
